Move Enform api route registration to Api folder
It's better to keep all api routes registering in the same place
