+++ functions.php
@@ -117,6 +117,7 @@ add_action(
         Rest::register_endpoints();
         Api\Gallery::register_endpoint();
         Api\Settings::register_endpoint();
+        Api\Enform::register_endpoints();
     }
 );
